chore: drop `scala_legacy` and `MainGenericRunner`
