--- compiler/test-coursier/run/envtest.scala
@@ -1,4 +1,0 @@
-package run
-
-object envtest extends App:
-  println("Hello " + sys.props("key"))
