--- compiler/test-coursier/run/myargs.txt
@@ -1,1 +1,0 @@
--classpath .
