--- compiler/test-coursier/run/myfile.scala
@@ -1,4 +1,0 @@
-package run
-
-object myfile extends App:
-  println("Hello")
