+++ compiler/test/dotty/tools/scripting/ClasspathTests.scala
@@ -78,44 +78,3 @@ class ClasspathTests:
 
       assert(hashbangClasspathJars.size == packlibJars.size)
   }
-  /*
-   * verify classpath is unglobbed by MainGenericRunner.
-   */
-  @Ignore
-  @Test def unglobClasspathVerifyTest = {
-    val testScriptName = "unglobClasspath_scalacli.sc"
-    val testScript = scripts("/scripting").find { _.name.matches(testScriptName) } match
-      case None => sys.error(s"test script not found: ${testScriptName}")
-      case Some(file) => file
-
-    val relpath = testScript.toPath.relpath.norm
-    val scalaCommand = scalaPath.relpath.norm
-    printf("===> unglobClasspathVerifyTest for script [%s]\n", relpath)
-    printf("bash is [%s]\n", bashExe)
-
-    if packBinScalaExists then
-      val sv = packScalaVersion
-      val tastyDirGlob = s"$packMavenDir/org/scala-lang/tasty-core_3/$sv/*"
-      // ^^^^^^^^^^^^^
-      // the classpath is a glob pattern that should be unglobbed by scala command,
-      // otherwise the script could not compile because it references a class
-      // from tasty-core
-
-      val bashCmdline = Seq(
-        "set +x ;",
-        "SCALA_OPTS=",
-        scalaCommand, "run", "--classpath", s"'$tastyDirGlob'", "--power", "--offline", "--server=false", relpath
-      ).mkString(" ")
-      val cmd = Array(bashExe, "-c", bashCmdline)
-
-      cmd.foreach { printf("[%s]\n", _) }
-
-      // test script reports the classpath it sees
-      val scriptOutput = exec(cmd*)
-      val scriptCp = findTaggedLine("unglobbed classpath", scriptOutput)
-      printf("%s\n", scriptCp)
-      val classpathJars = scriptCp.split(psep).map { _.getName }.sorted.distinct
-      //classpathJars.foreach { printf("%s\n", _) }
-      assert(classpathJars.size > 1)
-  }
-
