--- tests/run-with-compiler/i14541.check
@@ -1,2 +1,0 @@
-hello raw world
-hello run world
